feat: 심박수 데이터 wear os -> mobile 동기화 완료

See merge request s11-final/S11P31A408!38

+++ android/app/src/main/java/com/scare/ui/mobile/main/component/StressBar.kt
@@ -18,10 +18,8 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.onSizeChanged
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.text.font.FontWeight
@@ -35,7 +33,7 @@ import com.scare.ui.theme.low
 import com.scare.ui.theme.medium
 
 @Composable
-fun MyStressRate(stress: Int,
+fun MyStressRate(stress: Double,
                  modifier: Modifier = Modifier
 ) {
     val stressLevel = when {
